merge

+++ src/main/kotlin/com/example/cakezip/dto/EditOptionDto.kt
@@ -2,5 +2,5 @@ package com.example.cakezip.dto
 
 class EditOptionDto (
     val optionDetail:String = "",
-    val optionPrice:Long = 0
+    val optionPrice:Int = 0
     ){}
